requirement.txt
faiss-cpu
groq
PyPDF2
langchain
streamlit
langchain_community
python-dotenv
google-generativeai
pathlib
